+++ cmd/function_deploy.go
@@ -4,13 +4,13 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/vertigobr/safira/pkg/get"
-	s "github.com/vertigobr/safira/pkg/stack"
-	"os"
-
 	"github.com/vertigobr/safira/pkg/config"
 	d "github.com/vertigobr/safira/pkg/deploy"
 	"github.com/vertigobr/safira/pkg/execute"
+	"github.com/vertigobr/safira/pkg/get"
+	s "github.com/vertigobr/safira/pkg/stack"
+	"os"
+	"path/filepath"
 
 	"github.com/spf13/cobra"
 )
@@ -67,7 +67,8 @@ func runFunctionDeploy(cmd *cobra.Command, args []string) error {
 	functions := stack.Functions
 	if all {
 		for index, _ := range functions {
-			if err := checkDeployFiles(index, hostnameFlag); err!= nil {
+			swaggerFile := checkSwaggerFileExist(functions[index].Handler)
+			if err := checkDeployFiles(index, functions[index].Handler, hostnameFlag, swaggerFile); err!= nil {
 				return err
 			}
 
@@ -79,7 +80,9 @@ func runFunctionDeploy(cmd *cobra.Command, args []string) error {
 	} else {
 		for index, functionArg := range args {
 			if checkFunctionExists(args[index], functions) {
-				if err := checkDeployFiles(functionArg, hostnameFlag); err!= nil {
+				swaggerFile := checkSwaggerFileExist(functions[functionArg].Handler)
+
+				if err := checkDeployFiles(functionArg, functions[functionArg].Handler, hostnameFlag, swaggerFile); err!= nil {
 					return err
 				}
 
@@ -161,11 +164,12 @@ func deploy(kubectlPath, kubeconfigFlag, deployFolder, functionName string, verb
 	return nil
 }
 
-func checkDeployFiles(functionName, hostnameFlag string) error {
-	deployFolder := fmt.Sprintf("./deploy/%s/", functionName)
-	functionYaml := deployFolder + "function.yml"
-	ingressYaml  := deployFolder + "ingress.yml"
-	serviceYaml  := deployFolder + "service.yml"
+func checkDeployFiles(functionName, functionHandler, hostnameFlag , swaggerFile string) error {
+	deployFolder     := fmt.Sprintf("./deploy/%s/", functionName)
+	functionYamlName := deployFolder + "function.yml"
+	ingressYamlName  := deployFolder + "ingress.yml"
+	serviceYamlName  := deployFolder + "service.yml"
+	functionPath     := fmt.Sprintf("/function/%s", functionName)
 
 	if _, err := os.Stat(deployFolder); err != nil {
 		if err = os.MkdirAll(deployFolder, 0700); err != nil {
@@ -173,16 +177,76 @@ func checkDeployFiles(functionName, hostnameFlag string) error {
 		}
 	}
 
-	if err := d.CreateYamlFunction(functionYaml, functionName, functionsNamespace); err != nil {
+	var functionYaml d.K8sYaml
+	if err := functionYaml.MountFunction(functionName, functionsNamespace); err != nil {
 		return err
 	}
 
-	if err := d.CreateYamlIngress(ingressYaml, functionName, hostnameFlag); err != nil {
+	if err := functionYaml.CreateYamlFile(functionYamlName); err != nil {
 		return err
 	}
 
-	if err := d.CreateYamlService(serviceYaml, functionName, hostnameFlag); err != nil {
+	var ingressYaml d.K8sYaml
+	if err := ingressYaml.MountIngress(functionName, functionName, functionPath, hostnameFlag); err != nil {
 		return err
+	}
+
+	if err := ingressYaml.CreateYamlFile(ingressYamlName); err != nil {
+		return err
+	}
+
+	var serviceYaml d.K8sYaml
+	if err := serviceYaml.MountService(functionName, hostnameFlag, true); err != nil {
+		return err
+	}
+
+	if err := serviceYaml.CreateYamlFile(serviceYamlName); err != nil {
+		return err
+	}
+
+	if len(swaggerFile) > 1 {
+		deploymentSwaggerUIYaml := deployFolder + "swagger-ui-deployment.yml"
+		ingressSwaggerUIYaml    := deployFolder + "swagger-ui-ingress.yml"
+		serviceSwaggerUIYaml    := deployFolder + "swagger-ui-service.yml"
+		configMapSwaggerUIYaml  := deployFolder + "swagger-ui-config-map.yml"
+		swaggerUIName           := fmt.Sprintf("%s-swagger-ui", functionName)
+		swaggerPath             := fmt.Sprintf("/swagger-ui/%s", functionName)
+
+		var swaggerDeploymentYaml d.K8sYaml
+		if err := swaggerDeploymentYaml.MountDeployment(swaggerUIName, "swaggerapi/swagger-ui:v3.24.3", swaggerPath); err != nil {
+			return err
+		}
+
+		if err := swaggerDeploymentYaml.CreateYamlFile(deploymentSwaggerUIYaml); err != nil {
+			return err
+		}
+
+		var swaggerIngressYaml d.K8sYaml
+		if err := swaggerIngressYaml.MountIngress(swaggerUIName, swaggerUIName, swaggerPath, hostnameFlag); err != nil {
+			return err
+		}
+
+		if err := swaggerIngressYaml.CreateYamlFile(ingressSwaggerUIYaml); err != nil {
+			return err
+		}
+
+		var swaggerServiceYaml d.K8sYaml
+		if err := swaggerServiceYaml.MountService(swaggerUIName, hostnameFlag, false); err != nil {
+			return err
+		}
+
+		if err := swaggerServiceYaml.CreateYamlFile(serviceSwaggerUIYaml); err != nil {
+			return err
+		}
+
+		var swaggerConfigMapYaml d.K8sYaml
+		if err := swaggerConfigMapYaml.MountConfigMap(swaggerUIName, functionHandler + "/" + swaggerFile); err != nil {
+			return err
+		}
+
+		if err := swaggerConfigMapYaml.CreateYamlFile(configMapSwaggerUIYaml); err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -214,4 +278,18 @@ func rolloutFunction(kubectlPath, kubeconfig, functionName string, verboseFlag b
 	}
 
 	return nil
+}
+
+func checkSwaggerFileExist(handler string) string {
+	swaggerPath := filepath.Join(handler, "swagger.yml")
+	if _, err := os.Stat(swaggerPath); err == nil {
+		return "swagger.yml"
+	}
+
+	swaggerPath = filepath.Join(handler, "swagger.yaml")
+	if _, err := os.Stat(swaggerPath); err == nil {
+		return "swagger.yaml"
+	}
+
+	return ""
 }